// src/types/products.ts
export type Products = {
    dados: {
        id: string;
        nome: string;
        valor: string;
        id_someney_prod: string;
        id_pgs_produto: string;
        status: string;
        partner_id: number;
        tipo: string;
        subtipo: string;
        created_at: string
    }
} 